% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nv_naturvardsregistret.R
\name{get_nvrid_wkt}
\alias{get_nvrid_wkt}
\title{Title}
\usage{
get_nvrid_wkt(
  nvrid,
  beslutsstatus = "Gällande",
  crs = 3006,
  crs_to = NULL,
  map = FALSE,
  fill = TRUE,
  base_url = "https://geodata.naturvardsverket.se/naturvardsregistret/rest/v3/omrade"
)
}
\arguments{
\item{base_url}{}
}
\description{
Title
}
