% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_geografisk_indelning.R
\name{lm_geografisk_indelning}
\alias{lm_geografisk_indelning}
\title{Title}
\usage{
lm_geografisk_indelning(
  request,
  crs = 3006,
  rename = FALSE,
  format = c("sf", "json", "xml"),
  url = "https://api.lantmateriet.se/distribution/produkter/geografiskindelning/v2",
  token = httr::content(lm_get_token())$access_token
)
}
\arguments{
\item{token}{}
}
\description{
Title
}
