% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_api_token.R
\name{lm_get_token}
\alias{lm_get_token}
\title{Title}
\usage{
lm_get_token(
  url = "https://api.lantmateriet.se/token",
  consumer_key = Sys.getenv("consumer_key"),
  consumer_secret = Sys.getenv("consumer_secret")
)
}
\arguments{
\item{consumer_secret}{}
}
\description{
Title
}
