% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_hojdmodell.R
\name{lm_hojdmodell_download}
\alias{lm_hojdmodell_download}
\title{Title}
\usage{
lm_hojdmodell_download(
  .x,
  user = Sys.getenv("lm_atkomst_ver_user"),
  password = Sys.getenv("lm_atkomst_ver_pwd"),
  create_dirs = TRUE
)
}
\arguments{
\item{create_dirs}{}
}
\description{
Title
}
