% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_hojdmodell.R
\name{lm_hojdmodell_prepare}
\alias{lm_hojdmodell_prepare}
\title{Title}
\usage{
lm_hojdmodell_prepare(
  .x,
  ruta = .x$ruta_5,
  dx = 5000,
  dy = 5000,
  main_dir = NULL
)
}
\arguments{
\item{main_dir}{}
}
\description{
Title
}
