% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_hojdmodell.R
\name{lm_hojdmodell_url}
\alias{lm_hojdmodell_url}
\title{Title}
\usage{
lm_hojdmodell_url(
  crs,
  xmin,
  ymin,
  xmax,
  ymax,
  url_base = "https://download-ver.lantmateriet.se/hojdmodell/wcs/v1"
)
}
\arguments{
\item{url_base}{}
}
\description{
Title
}
