% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_markhojd_direkt.R
\name{lm_markhojd_geometry}
\alias{lm_markhojd_geometry}
\title{Title}
\usage{
lm_markhojd_geometry(
  geometry,
  encode = "raw",
  as = "text",
  url = "https://api.lantmateriet.se/distribution/produkter/hojd/v1/rest/api/hojd",
  token = httr::content(lm_get_token())$access_token
)
}
\arguments{
\item{token}{}
}
\description{
Title
}
