% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_markhojd_direkt.R
\name{lm_markhojd_point}
\alias{lm_markhojd_point}
\title{Title}
\usage{
lm_markhojd_point(
  east,
  north,
  altitude_only = FALSE,
  as = "text",
  url = "https://api.lantmateriet.se/distribution/produkter/hojd/v1/rest/api/hojd/3006",
  token = httr::content(lm_get_token())$access_token
)
}
\arguments{
\item{token}{}
}
\description{
Title
}
