% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_ortnamn.R
\name{lm_ortnamn_coords}
\alias{lm_ortnamn_coords}
\title{Title}
\usage{
lm_ortnamn_coords(
  easting,
  northing,
  crs = 3006,
  token = httr::content(lm_get_token())$access_token
)
}
\arguments{
\item{token}{}
}
\description{
Title
}
