% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_ortnamn.R
\name{lm_ortnamn_kriterier}
\alias{lm_ortnamn_kriterier}
\title{Title}
\usage{
lm_ortnamn_kriterier(
  query,
  as = "text",
  url = "https://api.lantmateriet.se/distribution/produkter/ortnamn/v2.1",
  token = httr::content(lm_get_token())$access_token
)
}
\arguments{
\item{token}{}
}
\description{
Title
}
