% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nv_naturvardsregistret.R
\name{nv_rest_api}
\alias{nv_rest_api}
\title{Title}
\usage{
nv_rest_api(
  str_parameters,
  ...,
  base_url = "https://geodata.naturvardsverket.se/naturvardsregistret/rest/v3",
  remove_atom_link = FALSE,
  convert_unix_timestamp = TRUE
)
}
\arguments{
\item{convert_unix_timestamp}{}
}
\description{
Title
}
