% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nv_naturvardsregistret.R
\name{nv_rest_api_httr}
\alias{nv_rest_api_httr}
\title{Title}
\usage{
nv_rest_api_httr(
  .x,
  base_url =
    "https://geodata.naturvardsverket.se/naturvardsregistret/rest/v3/omrade/nolinks"
)
}
\arguments{
\item{base_url}{}
}
\description{
Title
}
