% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_geografisk_indelning.R
\name{rename_geografisk_indelning}
\alias{rename_geografisk_indelning}
\title{Title}
\usage{
rename_geografisk_indelning(
  .data,
  .fix_names = c(detaljtyp = "typ", lankod = "lanskod", lanbok = "lansbokstav", lannamn =
    "lansnamn", geom = "geometry", geom = "Lansyta", geom = "Kommunyta")
)
}
\arguments{
\item{.fix_names}{}
}
\description{
Title
}
