% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_geografisk_indelning.R
\name{rename_geografisk_indelning_gml}
\alias{rename_geografisk_indelning_gml}
\title{Title}
\usage{
rename_geografisk_indelning_gml(
  .file,
  .detaljtyp,
  .fix_names = c(detaljtyp = "typ", lankod = "lanskod", lanbok = "lansbokstav", lannamn =
    "lansnamn", geom = "geometry", geom = "Lansyta", geom = "Kommunyta", id = "gml_id")
)
}
\arguments{
\item{.fix_names}{}
}
\description{
Title
}
